Sketch out TasProcessor and move _load_data.
