+++ indra/sources/tas/processor.py
@@ -3,18 +3,19 @@ from builtins import dict, str
 
 __all__ = []
 
-import csv
-from os import path
 
-HERE = path.dirname(path.abspath(__file__))
-DATAFILE_NAME = 'classification_hms_cmpds_symbol.csv'
+class TasProcessor(object):
+    """A processor for Target Affinity Spectrum data compiled by N. Moret.
 
+    This data was compiled in the HMS LSP as an improvement on the "arbitrary"
+    selection of targets present in the similar LINCS dataset.
+    """
+    def __init__(self, data):
+        self._data = data
+        self.statements = []
+        for row in data:
+            self._process_row(row)
+        return
 
-def _load_data():
-    csv_path = path.join(HERE, path.pardir, path.pardir, path.pardir, 'data',
-                         DATAFILE_NAME)
-    with open(csv_path, 'r') as f:
-        csv_lines = f.readlines()
-    reader = csv.reader(csv_lines)
-    headers = reader.__next__()
-    return [{headers[i]: val for i, val in enumerate(line)} for line in reader]
+    def _process_row(self, row):
+        pass
